Move the sidebar refreshes, viewport layout and entity actions out of App.vue
Three more clusters leave the shell: the debounced sidebar refreshes with
their timers, the viewport-driven layout (which rails fit, how many grid
columns fit beside them) and the post-action reconciliation for assigning,
moving, exporting and clearing the search.

App.vue is now 905 lines - 504 of script and 401 of template - down from
2455 at the start of the step.
